Add data description file and excels for auto production and self consumption
Use folder "electricity generation_main vs main activity" for autoproduction and "electricity_self consumption_distr losses_grid heat dem" for self consumption. Readmes are in the folders.

Autoproduction and slef consumption have to be deducted from the electricity demand given in tab 33 of the Balmorel input file. Additionally, the transmission and distribution losses (as given in the input file) have to be deducted in tab 33.